Cleanup and enhancements - Debug init in device - Init macros - Variable buffer count - Fixed easyhwnd warnings - Added function COMGetable - ComPointer naming (only for D3D12Objects) - Internal naming (including macros)

+++ ModernGXLib/src/ModernGX/Util/EasyHWND.h
@@ -374,7 +374,7 @@ namespace EasyHWND {
 			/// <param name="ptrHeight">Pointer to height</param>
 			void adjustSizeForWindow(UINT* ptrWidth, UINT* ptrHeight) {
 				// Create window rect
-				RECT windowRect = {0, 0, *ptrWidth, *ptrHeight};
+				RECT windowRect = {0, 0, (LONG)*ptrWidth, (LONG)*ptrHeight};
 
 				// Adjust window rect
 				AdjustWindowRectEx(&windowRect, m_style, m_bHasMenue, m_styleEx);
